web-animations/animation-player/animation-player-start-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer startTime attribute test</title>
<meta name="assert" content="Returns the start time of this player. Setting this attribute updates the player start time using the procedure to update the player start time of this object to the new value">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationplayer-starttime">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = newAnimationPlayer(createDiv(this));

    var newStartTime = document.timeline.currentTime + 100;
    player.startTime = newStartTime;

    assert_equals(player.startTime, newStartTime,
        'The value of AnimationPlayer.startTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.startTime attribute to new value');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.startTime = null;

    assert_equals(player.startTime, null,
        'The value of AnimationPlayer.startTime attribute should be set to the null value');
}, 'Test setting value of AnimationPlayer.startTime attribute to unresolved value');

test(function() {
    var player = new AnimationPlayer();
    var newStartTime = document.timeline.currentTime + 100;
    player.startTime = newStartTime;

    assert_equals(player.startTime, newStartTime,
        'The value of AnimationPlayer.startTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.startTime attribute to new value ' +
    'for animation player with no associated timeline');

test(function() {
    var player = new AnimationPlayer();
    player.currentTime = document.timeline.currentTime + 200;
    var newStartTime = document.timeline.currentTime + 100;
    player.startTime = newStartTime;

    assert_equals(player.currentTime, null,
        'Setting AnimationPlayer.startTime attribute for AnimationPlayer with no timeline ' +
        'should make current time unresolved');
}, 'Test that setting AnimationPlayer.startTime attribute for AnimationPlayer with no timeline ' +
    'makes current time unresolved');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.ready.then(t.step_func(function() {
        var newStartTime = document.timeline.currentTime + 50;
        player.startTime = newStartTime;
        // New player start time is in 50ms.
        // Wait for 100ms and check that currentTime is about 50ms
        setTimeout(t.step_func(function() {
            assert_approx_equals(player.currentTime, 50, TIME_EPSILON, 'The value of ' +
                'AnimationPlayer.currentTime should correspond to startTime');
            t.done();
        }), 100);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test setting AnimationPlayer start time in the future moves beginning of play ' +
    'accordingly');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));

    player.ready.then(t.step_func(function() {
        var newStartTime = document.timeline.currentTime - 100;
        player.startTime = newStartTime;
        setTimeout(t.step_func(function() {
            assert_approx_equals(player.currentTime, 150, TIME_EPSILON, 'The value of ' +
                'AnimationPlayer.currentTime should correspond to startTime');
            t.done();
        }), 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test setting AnimationPlayer start time in the past shifts beginning of play');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.ready.then(t.step_func(function() {
        var newStartTime = document.timeline.currentTime - ANIMATION_DURATION - 100;
        player.startTime = newStartTime;
        assert_equals(player.currentTime, ANIMATION_DURATION, 'The value of ' +
            'AnimationPlayer.currentTime should correspond updated startTime');
        assert_equals(player.playState, 'finished', 'Player should be finished');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test setting AnimationPlayer start time to the value in the past that ' +
    'earlier than current time minus animation duration finishes the player');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    var newStartTime = document.timeline.currentTime + 100;
    player.startTime = newStartTime;
    player.ready.then(t.step_func(function(plr) {
        assert_equals(plr, player, 'The current ready promise should be fulfilled with the player');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that if player has a pending play task, player\'s current ready promise ' +
    'is fulfilled with player');
</script>
</body>
